+++ src/objects/impls/structure_factory.rs
@@ -1,0 +1,7 @@
+use crate::{constants::ResourceType, constants::ReturnCode, objects::StructureFactory};
+
+impl StructureFactory {
+    pub fn produce(&self, ty: ResourceType) -> ReturnCode {
+        js_unwrap! {@{self.as_ref()}.produce(__resource_type_num_to_str(@{ty as u32}))}
+    }
+}