+++ src/objects/impls/structure_invader_core.rs
@@ -1,0 +1,8 @@
+use crate::objects::StructureInvaderCore;
+
+simple_accessors! {
+    impl StructureInvaderCore {
+        pub fn level() -> u32 = level;
+        pub fn ticks_to_deploy() -> Option<u32> = ticksToDeploy;
+    }
+}